requirements.txt
certifi==2019.9.11
python-chess==0.28.3
wincertstore==0.2
